% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{full_scenarios_core}
\alias{full_scenarios_core}
\title{Exported version of full_scenarios_core}
\usage{
full_scenarios_core(
  big_b,
  big_M,
  obs,
  path,
  shocks,
  h,
  n_var,
  g_ = NULL,
  Sigma_g_ = NULL
)
}
\arguments{
\item{big_b}{Cube of B matrices}

\item{big_M}{Cube of M matrices}

\item{obs}{Indices of constrained observables}

\item{path}{Flattened path for observables}

\item{shocks}{Indices of shocks to be recovered}

\item{h}{Forecast horizon}

\item{n_var}{Number of variables}

\item{g_}{Optional vector of non-driving shocks}

\item{Sigma_g_}{Optional covariance matrix of non-driving shocks}
}
\description{
This function wraps the Rcpp-exported version of \code{full_scenarios_core}
and allows external users to call it with correct argument checks.
}
