% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.r
\name{KL}
\alias{KL}
\title{KL function
APR suggest this measure to assess the "plausibility" of the conditional forecast.
It is based on the Kullback-Leibler measure of distance between the unconditional forecast and the
conditional/scenario forecast.}
\usage{
KL(Sigma_eps, mu_eps, h, plot_ = F, max_cores = NULL)
}
\arguments{
\item{Sigma_eps}{variance of innovation}

\item{mu_eps}{mean of innovation}

\item{h}{forecast horizon}

\item{plot_}{logical; if TRUE then a histogram of the KL measure is returned}

\item{max_cores}{maximum number of cores to use for parallel processing (default: NULL, uses CRAN-compliant detection with Windows=1)}
}
\value{
Returns the APR 'q': ie distance from a fair binomial distribution
}
\description{
KL function
APR suggest this measure to assess the "plausibility" of the conditional forecast.
It is based on the Kullback-Leibler measure of distance between the unconditional forecast and the
conditional/scenario forecast.
}
