% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NKdata.R
\docType{data}
\name{NKdata}
\alias{NKdata}
\title{Example Dataset NKdata}
\format{
A data frame with 244 rows and 4 variables:
\describe{
\item{GDP.E}{GDP}
\item{pi.PCE.core}{inflation}
\item{i}{interest rate}
\item{year}{year}
}
}
\usage{
data(NKdata)
}
\description{
A dataset used in the APRScenario package.
}
\keyword{datasets}
