% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bvars.r
\name{plot_bvars}
\alias{plot_bvars}
\title{plot_bvars: This function plots the IRFs generated with the BVAR}
\usage{
plot_bvars(
  M,
  significance_level = 0.05,
  central_tendency = "mean",
  variable_names = NULL,
  shock_names = NULL
)
}
\arguments{
\item{M}{IRFs produced by eg bvarSIGNs}

\item{significance_level}{(eg 0.05)}

\item{central_tendency}{eg 'mean' or 'median'}

\item{variable_names}{vector of names of variables (strings)}

\item{shock_names}{vector of names of variables (strings)}
}
\value{
a list of ggplot objects (plots)
}
\description{
plot_bvars: This function plots the IRFs generated with the BVAR
}
