% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cond_forc.r
\name{plot_cond_forc}
\alias{plot_cond_forc}
\title{plot_cond_forc function; Data should conatain the variable "variable", the "hor" horizon and a "history"}
\usage{
plot_cond_forc(variable = NULL, data = NULL)
}
\arguments{
\item{variable}{name of variable to be plotted (string)}

\item{data:}{conditional forecast data frame (eg cond.for)}
}
\value{
plot
}
\description{
plot_cond_forc function; Data should conatain the variable "variable", the "hor" horizon and a "history"
}
