% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{scenarios}
\alias{scenarios}
\title{scenarios function (fully optimized with Rcpp)
This function computes the mean and covariances to draw from the conditional forecast
The actual draw is done in the simscen function}
\usage{
scenarios(
  h = 3,
  path = NULL,
  obs = NULL,
  free_shocks = NULL,
  n_draws,
  n_sample = n_draws,
  n_var,
  n_p,
  data_ = Z,
  g = NULL,
  Sigma_g = NULL
)
}
\arguments{
\item{h}{forecast horizon}

\item{path}{conditional path of observables}

\item{obs}{position of observable(s)}

\item{free_shocks}{position of non-driving shocks (NA if all driving)}

\item{n_draws}{Number of posterior draws}

\item{n_sample}{Number of draws to sample (<= n_draws)}

\item{n_var}{Number of variables}

\item{n_p}{Number of lags}

\item{data_}{Optional matrix of data (default is Z)}

\item{g}{Optional matrix of non-driving shocks}

\item{Sigma_g}{Optional covariance matrix of non-driving shocks}
}
\value{
list of mu_eps, Sigma_eps, mu_y, Sigma_y, big_b, big_M, draws_used
}
\description{
scenarios function (fully optimized with Rcpp)
This function computes the mean and covariances to draw from the conditional forecast
The actual draw is done in the simscen function
}
