% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios_old.r
\name{scenarios_old}
\alias{scenarios_old}
\title{scenarios function}
\usage{
scenarios_old(
  h = 3,
  path = NULL,
  obs = NULL,
  shocks = NULL,
  n_draws,
  n_sample = n_draws,
  n_var,
  n_p,
  data_ = Z
)
}
\arguments{
\item{h}{forecast horizon}

\item{path}{conditional path of observables}

\item{obs}{position of observable(s)}

\item{shocks}{position of non-driving shocks (NA if all driving)}

\item{n_draws}{Number of draws}

\item{n_sample}{Number of draws to sample (must be <= n_draws)}

\item{n_var}{Number of variables}

\item{n_p}{Number of lags}

\item{data_}{(matrix optional) The data, stacking Y over X (default is Z produced by gen_mats)}
}
\value{
mean and variance of innovations and forecasts (four matrices) plus big_b and big_M
}
\description{
scenarios function
}
