% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_b_and_M.r
\name{big_b_and_M}
\alias{big_b_and_M}
\title{big_b_and_M
This function returns the extended b and M matrices as in APR}
\usage{
big_b_and_M(h, n_draws, n_var, n_p, data_ = Z)
}
\arguments{
\item{h}{forecast horison}

\item{n_draws}{Number of draws}

\item{n_var}{Number of variables}

\item{n_p}{Number of lags}

\item{data_}{(matrix optional) The data, stacking Y over X (data and laggs)
-- columns are observations (default is Z produced by gen_mats)
NB: this is not necessarily the same as the data used to estimate the model
If run counterfactuals in previoius historical period (ie not forecast) must pass the data up to previous period relative to counterfactual}
}
\value{
the big_b and big_M matrices of mean and IRF
}
\description{
big_b_and_M
This function returns the extended b and M matrices as in APR
}
