% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forc_h.r
\name{forc_h}
\alias{forc_h}
\title{forc_h function}
\usage{
forc_h(h = 1, n_sim = 200)
}
\arguments{
\item{h}{forecast horizon}

\item{n_sim}{length of shock simulation}
}
\value{
a matrix of unconditional forecasts
}
\description{
forc_h function
}
