% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mats.r
\name{gen_mats}
\alias{gen_mats}
\title{gen_mats function}
\usage{
gen_mats(posterior = NULL, n_draws, n_var, n_p)
}
\arguments{
\item{posterior}{Posterior estimation results (eg from BsvarSIGNs)}

\item{n_draws}{Number of draws}

\item{n_var}{Number of variables}

\item{n_p}{Number of lags}
}
\value{
Returns all objects necessary for scenario analysis (e.g., IRF matrix),
including: \code{M}, \code{M_inv}, \code{M_list}, \code{B}, \code{B_list},
\code{n_p}, \code{n_var}, \code{Y}, \code{X}, and \code{Z}.
}
\description{
this function assign to the global environment the matrices necessary for forecasts
}
