% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mats.r
\name{gen_mats}
\alias{gen_mats}
\title{gen_mats function}
\usage{
gen_mats()
}
\value{
Returns all the objects necessary for the other functions in scenario analysis (eg IRF matrix)
}
\description{
this function assign to the global environment the matrices necessary for forecasts
}
