% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cond_forc.r
\name{plot_cond_forc}
\alias{plot_cond_forc}
\title{plot_cond_forc function}
\usage{
plot_cond_forc(variable = varbls[1])
}
\arguments{
\item{variable}{name of variable to be plotted (string)}
}
\value{
plot
}
\description{
plot_cond_forc function
}
