% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.r
\name{scenarios}
\alias{scenarios}
\title{scenarios function}
\usage{
scenarios(h = 3, path = NULL, obs = NULL, shocks = NULL)
}
\arguments{
\item{h}{forecast horizon}

\item{path}{conditional path of observables}

\item{obs}{position of observable(s)}

\item{shocks}{position of non-driving shocks (NA if all driving)}
}
\value{
mean and variance of innovations and forecasts (four matrices) plus big_b and big_M
}
\description{
scenarios function
}
