% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simscen.r
\name{SimScen}
\alias{SimScen}
\title{simscen function (optimized)}
\usage{
SimScen(
  mu_eps,
  Sigma_eps,
  mu_y,
  Sigma_y,
  big_b,
  big_M,
  n_sim,
  h,
  varbls,
  idx_sampled = 1:dim(mu_eps)[3]
)
}
\arguments{
\item{mu_eps}{mean innovation}

\item{Sigma_eps}{variance innovation}

\item{mu_y}{mean forecast}

\item{Sigma_y}{variance forecast}

\item{big_b}{history forecast}

\item{big_M}{IRF (innovation loading)}

\item{n_sim}{number of simulations}

\item{h}{horizon}

\item{varbls}{variable names}

\item{idx_sampled}{index of random sample to use instead of full draws (from scenarios)}
}
\value{
conditional forecast path and distribution
}
\description{
simscen function (optimized)
}
