% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simscen.r
\name{SimScen}
\alias{SimScen}
\title{simscen function}
\usage{
SimScen(mu_eps, Sigma_eps, mu_y, Sigma_y, big_b, big_M, n_sim, h, varbls)
}
\arguments{
\item{mu_eps}{mean innovation}

\item{Sigma_eps}{variance innovation}

\item{mu_y}{mean forecast}

\item{Sigma_y}{variance forecast}

\item{big_b}{history forecast}

\item{big_M}{IRF (innovation loading)}

\item{n_sim}{number of simulations}

\item{h}{horizon}
}
\value{
conditional forecast path and distribution
}
\description{
simscen function
}
